gm8exe/asset/{...}: u32 as u32 (wtf)

+++ gm8exe/src/asset/script.rs
@@ -32,7 +32,7 @@ impl Asset for Script {
 
     fn serialize_exe(&self, mut writer: impl io::Write, version: GameVersion) -> io::Result<()> {
         writer.write_pas_string(&self.name)?;
-        writer.write_u32::<LE>(VERSION as u32)?;
+        writer.write_u32::<LE>(VERSION)?;
         writer.write_pas_string(&self.source)?;
         Ok(())
     }
